+++ packages/offscreen-display/CHANGELOG.md
@@ -7,9 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2024-12-12
+
 ### Added
 
 - introduce CHANGELOG.md ;)
+
+### Changed
+
+- upgrade dependencies
+  - `@spearwolf/eventize` to `4.x`
+  - `@spearwolf/signalize` to `0.18.x`
 
 
 ## [0.1.2] - 2024-02-27