+++ packages/rainbow-line/CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ...
 
+## [0.4.0] - 2024-12-12
+
+### Changed
+
+- upgrade dependencies
+  - `@spearwolf/eventize` to `^4.x`
+  - `@spearwolf/offscreen-display` to `^0.3.0`
+
 ## [0.3.0] - 2024-09-11
 
 ### Added